requirements.txt
dj-database-url==0.3.0
Django==1.8.1
django-postgrespool==0.3.0
psycopg2==2.6
SQLAlchemy==1.0.4
whitenoise==1.0.6
gunicorn==19.3.0
foursquare===1!2015.5.26
Pillow==2.7.0
virtualenv==13.0.3
requests==2.7.0
python-dateutil==2.4.2